Add vendor/ directory with .gitkeep file
At present the deployment process is trying to symlink a bundle
directory into this dir on deploy, but as vendor/ does not exist the
process fails, meaning the bundle is not cached.

This should allow the bundle to be cached between deploys, which should
greatly speed up the deployment process.